Update plugin system
