+++ web/src/pluginSystem/index.ts
@@ -1,1 +1,2 @@
+export * from "./pluginSystem_legacy";
 export * from "./pluginSystem";
